# docs/swagger.yaml
definitions:
  helper.Response:
    properties:
      data: {}
      description: {}
      isError:
        type: boolean
      status:
        type: string
    type: object
  model.Address:
    properties:
      city:
        type: string
      country:
        type: string
      id:
        type: integer
      is_primary:
        type: boolean
      orders:
        items:
          $ref: '#/definitions/model.Order'
        type: array
      phone_number:
        type: string
      receiver_name:
        type: string
      user:
        $ref: '#/definitions/model.User'
      user_id:
        type: integer
    type: object
  model.Bank:
    properties:
      bank_code:
        type: string
      bank_name:
        type: string
      id:
        type: integer
      image:
        type: string
    required:
    - bank_code
    - bank_name
    type: object
  model.Category:
    properties:
      category:
        type: string
      id:
        type: integer
      products:
        items:
          $ref: '#/definitions/model.Product'
        type: array
    type: object
  model.CreditCard:
    properties:
      card_number:
        type: string
      expired_month:
        type: integer
      expired_year:
        type: integer
      id:
        type: integer
      orders:
        items:
          $ref: '#/definitions/model.Order'
        type: array
      user_id:
        type: integer
    type: object
  model.Order:
    properties:
      address_id:
        type: integer
      amount:
        type: number
      created_at:
        type: string
      credit_card_id:
        type: integer
      gopay_id:
        type: integer
      id:
        type: string
      order_status:
        type: string
      products:
        items:
          $ref: '#/definitions/model.Product'
        type: array
      transaction_bank_id:
        type: integer
      transactions:
        items:
          $ref: '#/definitions/model.Transaction'
        type: array
      updated_at:
        type: string
      user_id:
        type: integer
    type: object
  model.Product:
    properties:
      brand_id:
        type: integer
      cart_users:
        items:
          $ref: '#/definitions/model.User'
        type: array
      categories:
        description: many 2 many
        items:
          $ref: '#/definitions/model.Category'
        type: array
      category_id:
        type: integer
      created_at:
        type: string
      description:
        type: string
      id:
        type: integer
      image_url:
        type: string
      name:
        type: string
      orders:
        items:
          $ref: '#/definitions/model.Order'
        type: array
      price:
        type: number
      reviews:
        items:
          $ref: '#/definitions/model.Review'
        type: array
      size:
        items:
          $ref: '#/definitions/model.Size'
        type: array
      updated_at:
        type: string
      wishlisted_users:
        items:
          $ref: '#/definitions/model.User'
        type: array
    type: object
  model.ProductReview:
    properties:
      comment:
        type: string
      created_at:
        type: string
      full_name:
        type: string
      id:
        type: integer
      rating:
        type: number
    type: object
  model.Review:
    properties:
      comment:
        type: string
      created_at:
        type: string
      id:
        type: integer
      product_id:
        type: integer
      rating:
        type: number
      user_id:
        type: integer
    type: object
  model.Size:
    properties:
      id:
        type: integer
      product:
        items:
          $ref: '#/definitions/model.Product'
        type: array
      size:
        type: string
    type: object
  model.Transaction:
    properties:
      currency:
        type: string
      expiry_time:
        type: integer
      fraud_status:
        type: string
      gross_amount:
        type: integer
      id:
        type: string
      order_id:
        type: integer
      payment_type:
        type: string
      settlement_time:
        type: integer
      signature:
        type: string
      transaction_status:
        type: string
      transaction_time:
        type: integer
    type: object
  model.User:
    properties:
      addresses:
        items:
          $ref: '#/definitions/model.Address'
        type: array
      cart_products:
        items:
          $ref: '#/definitions/model.Product'
        type: array
      created_at:
        type: string
      credit_cards:
        items:
          $ref: '#/definitions/model.CreditCard'
        type: array
      email:
        type: string
      full_name:
        type: string
      id:
        type: integer
      image_url:
        type: string
      is_admin:
        type: boolean
      is_verified:
        type: boolean
      orders:
        items:
          $ref: '#/definitions/model.Order'
        type: array
      password:
        type: string
      reviews:
        items:
          $ref: '#/definitions/model.Review'
        type: array
      updated_at:
        type: string
      username:
        type: string
      verification_codes:
        items:
          $ref: '#/definitions/model.VerificationCode'
        type: array
      verification_tokens:
        items:
          $ref: '#/definitions/model.VerificationToken'
        type: array
      wishlisted_product:
        items:
          $ref: '#/definitions/model.Product'
        type: array
    type: object
  model.VerificationCode:
    properties:
      code:
        type: string
      expiry_date:
        type: string
      id:
        type: integer
      user_id:
        type: integer
    type: object
  model.VerificationToken:
    properties:
      expiry_date:
        type: string
      id:
        type: integer
      token:
        type: string
      user_id:
        type: integer
    type: object
  requests.AddressRequest:
    properties:
      city:
        type: string
      country:
        type: string
      is_primary:
        type: boolean
      phone_number:
        maxLength: 20
        type: string
      receiver_name:
        type: string
    required:
    - city
    - country
    - is_primary
    - phone_number
    - receiver_name
    type: object
...
host: localhost:8080
info:
  contact:
    email: soberkoder@swagger.io
    name: API Support
  description: This is a Final Project
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Laza
  version: "1.0"
paths:
...
  /address/{id}:
    delete:
      consumes:
      - application/json
      description: Delete the address corresponding to the input Id
      parameters:
      - description: ID of the address
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/helper.Response'
            - properties:
                code:
                  type: string
                data:
                  $ref: '#/definitions/model.Address'
                isError:
                  type: boolean
                status:
                  type: string
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/helper.Response'
            - properties:
                code:
                  type: integer
                description:
                  type: string
                isError:
                  type: boolean
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/helper.Response'
            - properties:
                code:
                  type: integer
                description:
                  type: string
                isError:
                  type: boolean
              type: object
      security:
      - JWT: []
      summary: Delete address identified by the given id
      tags:
      - address
    get:
      consumes:
      - application/json
      description: Get details of address corresponding is the input Id
      parameters:
      - description: ID of the address
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/helper.Response'
            - properties:
                code:
                  type: string
                data:
                  $ref: '#/definitions/model.Address'
                isError:
                  type: boolean
                status:
                  type: string
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/helper.Response'
            - properties:
                code:
                  type: integer
                description:
                  type: string
                isError:
                  type: boolean
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/helper.Response'
            - properties:
                code:
                  type: integer
                description:
                  type: string
                isError:
                  type: boolean
              type: object
      security:
      - JWT: []
      summary: Get Details for a given id
      tags:
      - address
    put:
      consumes:
      - application/json
      description: put details of address
      parameters:
      - description: update address
        in: body
        name: address
        required: true
        schema:
          $ref: '#/definitions/requests.AddressRequest'
      - description: ID of the address
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/helper.Response'
            - properties:
                code:
                  type: string
                data:
                  $ref: '#/definitions/model.Address'
                isError:
                  type: boolean
                status:
                  type: string
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/helper.Response'
            - properties:
                code:
                  type: integer
                description:
                  type: string
                isError:
                  type: boolean
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/helper.Response'
            - properties:
                code:
                  type: integer
                description:
                  type: string
                isError:
                  type: boolean
              type: object
      security:
      - JWT: []
      summary: put Details
      tags:
      - address
...
  /bank/{id}:
    delete:
      consumes:
      - application/json
      description: Delete details of bank corresponding is the input Id
      parameters:
      - description: ID of the bank
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/helper.Response'
            - properties:
                code:
                  type: string
                data:
                  $ref: '#/definitions/model.Bank'
                isError:
                  type: boolean
                status:
                  type: string
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/helper.Response'
            - properties:
                code:
                  type: integer
                description:
                  type: string
                isError:
                  type: boolean
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/helper.Response'
            - properties:
                code:
                  type: integer
                description:
                  type: string
                isError:
                  type: boolean
              type: object
      security:
      - JWT: []
      summary: Delete Details for a given id
      tags:
      - bank
    get:
      consumes:
      - application/json
      description: get the bank corresponding to the input Id
      parameters:
      - description: ID of the bank
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/helper.Response'
            - properties:
                code:
                  type: string
                data:
                  $ref: '#/definitions/model.Bank'
                isError:
                  type: boolean
                status:
                  type: string
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/helper.Response'
            - properties:
                code:
                  type: integer
                description:
                  type: string
                isError:
                  type: boolean
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/helper.Response'
            - properties:
                code:
                  type: integer
                description:
                  type: string
                isError:
                  type: boolean
              type: object
      security:
      - JWT: []
      summary: Get bank identified by the given id
      tags:
      - bank
    put:
      consumes:
      - multipart/form-data
      description: put details of address
      parameters:
      - description: ID of the bank
        in: path
        name: id
        required: true
        type: integer
      - in: formData
        name: bank_code
        required: true
        type: string
      - in: formData
        name: bank_name
        required: true
        type: string
      - description: bank
        in: formData
        name: image
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/helper.Response'
            - properties:
                code:
                  type: string
                data:
                  $ref: '#/definitions/model.Bank'
                isError:
                  type: boolean
                status:
                  type: string
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/helper.Response'
            - properties:
                code:
                  type: integer
                description:
                  type: string
                isError:
                  type: boolean
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/helper.Response'
            - properties:
                code:
                  type: integer
                description:
                  type: string
                isError:
                  type: boolean
              type: object
      security:
      - JWT: []
      summary: put Details
      tags:
      - bank
...
  /products/{id}/reviews/:
    get:
      consumes:
      - application/json
      description: get the bank corresponding to the input Id
      parameters:
      - description: ID of the Product
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/helper.Response'
            - properties:
                data:
                  $ref: '#/definitions/model.ProductReview'
                isError:
                  type: boolean
                status:
                  type: string
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/helper.Response'
            - properties:
                code:
                  type: integer
                description:
                  type: string
                isError:
                  type: boolean
              type: object
        "500":
          description: Internal Server Error
          schema:
            allOf:
            - $ref: '#/definitions/helper.Response'
            - properties:
                code:
                  type: integer
                description:
                  type: string
                isError:
                  type: boolean
              type: object
      security:
      - JWT: []
      summary: Get bank identified by the given id
      tags:
      - ShowReviewLimit
...
securityDefinitions:
  JWT:
    description: 'How to input in swagger : ''Bearer <insert_your_token_here>'''
    in: header
    name: X-Auth-Token
    type: apiKey
swagger: "2.0"